+++ src/sim/singles/wrist_long.h
@@ -36,7 +36,7 @@ class WristLong : public Wrist
 public:
      
     /// constructor
-    WristLong(SingleProp const*, Mecable const*, size_t point);
+    WristLong(SingleProp const*, Mecable const*, unsigned point);
 
     /// destructor
     ~WristLong();
